build_kernel: switch to clang 19.0.1

+++ kernel/kernel_device_modules-6.6/drivers/usb/mtu3/mtu3_core.c
@@ -1010,7 +1010,7 @@ static irqreturn_t mtu3_link_isr(struct mtu3 *mtu)
 		pm_runtime_get(mtu->dev);
 		mtu3_ep0_setup(mtu);
 
-		if (udev_speed >= MTU3_SPEED_SUPER && !mtu->bypass_manual_pu)
+		if ((int)udev_speed >= (int)MTU3_SPEED_SUPER && !mtu->bypass_manual_pu)
 			ssusb_phy_dp_pullup(mtu->ssusb);
 	}
